Fix(post): tags projection and response validation fix
- Align the post tags contract across backend and frontend by fixing native query projections for taggedUsername and createdAt, and by allowing nullable error/message fields in the frontend response schema.

- This prevents tag fetch failures in Docker and restores tag loading for posts with and without tags.

+++ backend/src/main/java/it/evodev/instagram/posts/dto/response/PostTagDTO.java
@@ -12,8 +12,8 @@ package it.evodev.instagram.posts.dto.response;
  */
 public record PostTagDTO(
         String taggedUsername,
-        Integer xPosition,
-        Integer yPosition,
+        Double xPosition,
+        Double yPosition,
         String createdAt
 ) {
 }
